feat(web): Move the price per dates to header row on joiner activity calendar
