+++ apps/web/module/Hosting/Reservations/Calendar/activity/joiner-calendar-table/activity-rows.tsx
@@ -4,14 +4,22 @@ import { ChevronDown, ChevronRight } from "lucide-react"
 import { useCalendarStore } from "../stores/use-joiner-store"
 import ActivityTimeSlotRows from "./activity-time-slot-rows"
 import { T_Calendar_Joiner_Activity, T_Joiner_Activity } from "@repo/contract"
+import { pricePerDatesJoiner, pricePerDatesPrivate } from "../helpers/price-per-dates"
 
 const ActivityRows = () => {
-  const { startDate, daysPerPage, collapsed, activityData, setCollapsed } =
+  const { startDate, daysPerPage, collapsed, activityData, setCollapsed, setIsEditPricePerDatesModalOpen,  setSelectedDate,
+    setSelectedUnitId } =
     useCalendarStore((state) => state)
   const toggleCollapse = (category: string) => {
     setCollapsed({ ...collapsed, [category]: !collapsed[category] })
   }
 
+  const handleOpenPricePerDatesModal = (date: string, category: string) => {
+    setIsEditPricePerDatesModalOpen(true)
+    setSelectedDate(date)
+    setSelectedUnitId(category)
+  }
+ 
   return (
     <>
       {activityData.map((activity: T_Calendar_Joiner_Activity, index) => (
@@ -31,12 +39,48 @@ const ActivityRows = () => {
               const date = format(addDays(startDate, i), "yyyy-MM-dd")
               const isToday =
                 format(date, "yyyy-MM-dd") === format(today, "yyyy-MM-dd")
+              const noReservationCount = activity.joinerActivities.reduce(
+                (count: any, joinerActivity: any) => {
+             
+                  const reservations = joinerActivity.reservations || [];
+                  const hasReservation = reservations.some(
+                    (reservation: any) => {
+                      const reservationStart = format(
+                        new Date(reservation.startDate),
+                        "yyyy-MM-dd"
+                      )
+                      const reservationEnd = format(
+                        new Date(reservation.endDate),
+                        "yyyy-MM-dd"
+                      )
+                      return date >= reservationStart && date <= reservationEnd
+                    }
+                  )
+                  return count + (hasReservation ? 0 : 1)
+                },
+                0
+              )
+              
               return (
                 <td
                   key={i}
-                  className={`border gap-1 text-sm p-2 h-max text-center text-gray-500 bg-gray-100 font-semibold max-w-24 ${i + 1 === daysPerPage && "border-r-0"} ${isToday && "bg-primary-100"}`}
+                  className={`border gap-1 hover:bg-gray-200 text-sm p-2 h-max text-center text-gray-500 font-semibold max-w-24 ${i + 1 === daysPerPage && "border-r-0"} ${isToday && "bg-primary-100"}`}
                 >
-                  <div className="flex flex-col"></div>
+                  <div
+                    onClick={(e) => {
+                      handleOpenPricePerDatesModal(date, activity.id)
+                      e.stopPropagation()
+                    }}
+                    className="flex flex-col"
+                  >
+                    <div>{noReservationCount}</div>
+                
+                      <div>
+                        &#8369;{pricePerDatesJoiner({ joinerActivity: activity.joinerActivities[0]!, date })}
+                      </div>
+                    
+                   
+                  </div>
                 </td>
               )
             })}
