+++ apps/web/module/Hosting/Reservations/Calendar/activity/joiner-calendar-table/modals/edit-price-per-dates-modal.tsx
@@ -7,7 +7,8 @@ import { useForm } from "react-hook-form"
 import { useQueryClient } from "@tanstack/react-query"
 import toast from "react-hot-toast"
 import { useCalendarStore } from "../../stores/use-joiner-store"
-import useUpdateUnitPricePerDate from "../../../hooks/useUpdaateUnitPricePerDate"
+import useUpdateActivityPricePerDate from "../../hooks/use-update-activity-price-per-date"
+import { QK_CALENDAR_JOINER_ACTIVITIES } from "../../constants"
 
 const EditPricePerDatesModal = () => {
   const {
@@ -21,7 +22,7 @@ const EditPricePerDatesModal = () => {
   const [dateFrom, setDateFrom] = useState(selectedDate || "")
   const queryClient = useQueryClient()
   const { handleSubmit } = useForm<T_Property>({})
-  const { mutate, isPending } = useUpdateUnitPricePerDate(
+  const { mutate, isPending } = useUpdateActivityPricePerDate(
     selectedUnitId as string
   )
 
@@ -39,13 +40,13 @@ const EditPricePerDatesModal = () => {
     const payload = {
       fromDate: dateFrom,
       toDate: toDate,
-      baseRate: Number(baseRate),
+      price: Number(baseRate),
     }
     mutate(payload, {
       onSuccess: (data: any) => {
         if (!data.error) {
           queryClient.invalidateQueries({
-            queryKey: ["calendar-property"],
+            queryKey: [QK_CALENDAR_JOINER_ACTIVITIES],
           })
           toast.success(data.message)
           setIsEditPricePerDatesModalOpen(false)
